refactor: Dialog component naming conventions and update project-wide imports.
