+++ src/components/Dialog/delete-dialog/index.jsx
