+++ src/components/Dialog/see-detail-dialog/index.jsx
