+++ src/pages/admin-page/shared/branch-management/shared/branch-add-dialog.jsx
@@ -1,4 +1,4 @@
-import AddDialog from "@/components/Dialog/Add_dialog";
+import AddDialog from "@/components/dialog/add-dialog";
 import { TextField } from "@mui/material";
 
 const BranchAddDialog = ({
