+++ src/pages/admin-page/shared/branch-management/shared/branch-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const BranchDeleteDialog = ({ open, onClose, onConfirm }) => {
   return (
