+++ src/pages/admin-page/shared/branch-management/shared/branch-edit-dialog.jsx
@@ -1,4 +1,4 @@
-import EditDialog from "@/components/Dialog/Edit_Dialog";
+import EditDialog from "@/components/dialog/edit-dialog";
 import { TextField } from "@mui/material";
 
 const BranchEditDialog = ({
