+++ src/pages/admin-page/shared/branch-management/shared/branch-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const BranchRestoreDialog = ({ open, onClose, onConfirm }) => {
   return (
