+++ src/pages/admin-page/shared/category-management/shared/category-add-dialog.jsx
@@ -1,4 +1,4 @@
-import AddDialog from "@/components/Dialog/Add_dialog";
+import AddDialog from "@/components/dialog/add-dialog";
 import {
   FormControl,
   InputLabel,
