+++ src/pages/admin-page/shared/category-management/shared/category-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const CategoryDeleteDialog = ({ open, onClose, onConfirm }) => {
   return (
