+++ src/pages/admin-page/shared/category-management/shared/category-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const CategoryRestoreDialog = ({ open, onClose, onConfirm }) => {
   return (
