+++ src/pages/admin-page/shared/color-management/shared/color-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const ColorDeleteDialog = ({ open, onClose, onConfirm }) => {
   return (
