+++ src/pages/admin-page/shared/file-management/shared/file-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const FileDeleteDialog = ({ open, onClose, onConfirm }) => {
   return (
