+++ src/pages/admin-page/shared/product-management/shared/product-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const ProductDeleteDialog = ({ open, onClose, onConfirm }) => (
   <DeleteDialog
