+++ src/pages/admin-page/shared/product-management/shared/product-edit-dialog.jsx
@@ -5,7 +5,7 @@ import {
   Select,
   MenuItem,
 } from "@mui/material";
-import EditDialog from "@/components/Dialog/Edit_Dialog";
+import EditDialog from "@/components/dialog/edit-dialog";
 
 const ProductEditDialog = ({
   open,
