+++ src/pages/admin-page/shared/product-management/shared/product-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const ProductRestoreDialog = ({ open, onClose, onConfirm }) => (
   <RestoreDialog
