+++ src/pages/admin-page/shared/product-variant-management/shared/product-variant-edit-dialog.jsx
@@ -1,5 +1,5 @@
 import { TextField } from "@mui/material";
-import EditDialog from "@/components/Dialog/Edit_Dialog";
+import EditDialog from "@/components/dialog/edit-dialog";
 
 const ProductVariantEditDialog = ({
   openEditDialog,
