+++ src/pages/admin-page/shared/product-variant-management/shared/product-variant-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const ProductVariantRestoreDialog = ({ open, onClose, onConfirm }) => (
   <RestoreDialog
