+++ src/pages/admin-page/shared/promotion-management/shared/promotion-delete-dialog.jsx
@@ -1,4 +1,4 @@
-import DeleteDialog from "@/components/Dialog/Delete_dialog";
+import DeleteDialog from "@/components/dialog/delete-dialog";
 
 const PromotionDeleteDialog = ({ open, onClose, onConfirm }) => {
   return (
