+++ src/pages/admin-page/shared/promotion-management/shared/promotion-edit-dialog.jsx
@@ -1,4 +1,4 @@
-import EditDialog from "@/components/Dialog/Edit_Dialog";
+import EditDialog from "@/components/dialog/edit-dialog";
 import { Box, TextField } from "@mui/material";
 import { DatePicker, LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDayjs } from "@mui/x-date-pickers/AdapterDayjs";
