+++ src/pages/admin-page/shared/promotion-management/shared/promotion-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const PromotionRestoreDialog = ({ open, onClose, onConfirm }) => {
   return (
