+++ src/pages/admin-page/shared/user-management/shared/user-restore-dialog.jsx
@@ -1,4 +1,4 @@
-import RestoreDialog from "@/components/Dialog/Restore_dialog";
+import RestoreDialog from "@/components/dialog/restore-dialog";
 
 const UserRestoreDialog = ({ open, onClose, onConfirm, title, children }) => {
   return (
