+++ src/services/api/index.js
@@ -70,9 +70,6 @@ export const axiosBaseQuery =
         localStorage.removeItem("refreshToken");
         localStorage.removeItem("persist:root");
 
-        // Redirect to login page
-        window.location.replace("/");
-
         return { error };
       }
